pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resbot"
version = "0.1.0"
description = "Speed-optimized restaurant reservation bot for Resy and OpenTable"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "orjson>=3.9",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "sse-starlette>=2.0",
]

[project.optional-dependencies]
opentable = ["playwright>=1.40"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21", "ruff>=0.3"]

[project.scripts]
resbot = "resbot.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
